+++ src/http.rs
@@ -4,11 +4,15 @@ use actix_web::{
     error::{Error, JsonPayloadError},
     http::StatusCode,
     middleware::{ErrorHandlerResponse, ErrorHandlers},
+    web::{Form, Json},
+    Either,
     HttpRequest,
 };
 use serde_json::json;
 
 use crate::errors::HttpError;
+
+pub type FormOrJson<T> = Either<Form<T>, Json<T>>;
 
 /// Error handler for 401 Unauthorized
 pub fn create_auth_error_handler<B: MessageBody + 'static>() -> ErrorHandlers<B> {
@@ -30,6 +34,7 @@ pub fn create_auth_error_handler<B: MessageBody + 'static>() -> ErrorHandlers<B>
         })
 }
 
+/// Convert JSON payload deserialization errors into validation errors
 pub fn json_error_handler(
     error: JsonPayloadError,
     _: &HttpRequest,